requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
openpyxl>=3.1.0
plotly>=5.18.0
wordcloud>=1.9.0
matplotlib>=3.8.0
python-pptx>=1.0.0
anthropic>=0.30.0
PyMuPDF>=1.24.0
